+++ changelog.md
@@ -1,20 +1,44 @@
+### v.0.1.2
+- add controllers
+>- order
+
+- refine controllers
+>- category
+>- product
+
+- refine migrations
+>- create-order
+>- create-product-order
+>- create-product
+
+- add routes
+>- orders
+
+- refine models
+>- order
+>- product
+>- productorder
+
+- add seeders
+>- products
+
 ### v.0.1.1
 - add controllers
--- category
--- product
+>- category
+>- product
 - refine migrations
--- create-product
--- create-categories
+>- create-product
+>- create-categories
 - add routes
--- categories
--- products
+>- categories
+>- products
 - refine models
--- categories
--- product
--- productOrder
+>- categories
+>- product
+>- productOrder
 - add changelog
 - add seeders
--- categories
+>- categories
 
 ### v.0.1.0
 init project